requirements.txt
# NeuralNote Backend Dependencies

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.3.0

# AI/ML Services
openai==1.12.0
httpx==0.26.0

# Data Processing
numpy==1.26.3
pandas==2.1.4

# File Storage (Optional - for OSS)
oss2==2.18.4

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.32.5
aiofiles==23.2.1

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
black==23.12.1
isort==5.13.2
flake8==6.1.0
mypy==1.7.1

# Vector operations (for PgVector)
pgvector==0.2.4
